chore(repo): jest heap usage (#12706)
* chore(repo): use --logHeapUsage with jest

* chore(datastore): break down unit tests to avoid heap out of memory
