+++ packages/api-graphql/package.json
@@ -14,7 +14,7 @@
 		"access": "public"
 	},
 	"scripts": {
-		"test": "npm run lint && jest -w 1 --coverage",
+		"test": "npm run lint && jest -w 1 --coverage --logHeapUsage",
 		"test:watch": "tslint 'src/**/*.ts' && jest -w 1 --watch",
 		"build-with-test": "npm test && npm build",
 		"build:umd": "webpack && webpack --config ./webpack.config.dev.js",
