+++ packages/auth/package.json
@@ -14,7 +14,7 @@
 		"access": "public"
 	},
 	"scripts": {
-		"test": "yarn lint --fix && jest -w 1 --coverage",
+		"test": "yarn lint --fix && jest -w 1 --coverage --logHeapUsage",
 		"build-with-test": "npm test && npm run build",
 		"build:umd": "webpack && webpack --config ./webpack.config.dev.js",
 		"build:esm-cjs": "rollup -c rollup.config.mjs",
