+++ packages/geo/package.json
@@ -14,7 +14,7 @@
 		"./dist/esm/geo/geo.mjs"
 	],
 	"scripts": {
-		"test": "yarn run lint && jest -w 1 --coverage",
+		"test": "yarn run lint && jest -w 1 --coverage --logHeapUsage",
 		"test:size": "size-limit",
 		"build-with-test": "npm test && npm run build",
 		"build:umd": "webpack && webpack --config ./webpack.config.dev.js",
